main: added some comments for clarity and cleaned up some code.

+++ optimisation/ncnn_conversion_custom_models.py
@@ -7,4 +7,4 @@ model = YOLO("/home/dys/pi/intelligent_bird_scarer/models/my_fifth_trained_model
 # imgsz: Should match the size used during training (usually 640)
 # half:  Uses FP16 precision (highly recommended for Pi 5 to boost FPS)
 # int8:  Optional, but half=True is usually the "sweet spot" for Pi 5
-model.export(format="ncnn", imgsz=640, int8=True)
+model.export(format="ncnn", imgsz=640, half=True)
